+++ samples/README.md
@@ -35,19 +35,26 @@ __Usage:__ `node recognize.js --help`
 recognize.js <command>
 
 Commands:
-  recognize.js sync <filename>           Detects speech in a local audio file.
-  recognize.js sync-gcs <gcsUri>         Detects speech in an audio file located in a Google Cloud Storage bucket.
-  recognize.js sync-words <filename>     Detects speech in a local audio file with word time offset.
-  recognize.js async <filename>          Creates a job to detect speech in a local audio file, and waits for the job to
-                                         complete.
-  recognize.js async-gcs <gcsUri>        Creates a job to detect speech in an audio file located in a Google Cloud
-                                         Storage bucket, and waits for the job to complete.
-  recognize.js async-gcs-words <gcsUri>  Creates a job to detect speech  with word time offset in an audio file located
-                                         in a Google Cloud Storage bucket, and waits for the job to complete.
-  recognize.js stream <filename>         Detects speech in a local audio file by streaming it to the Speech API.
-  recognize.js listen                    Detects speech in a microphone input stream. This command requires that you
-                                         have SoX installed and available in your $PATH. See
-                                         https://www.npmjs.com/package/node-record-lpcm16#dependencies
+  recognize.js sync <filename>                   Detects speech in a local audio file.
+  recognize.js sync-gcs <gcsUri>                 Detects speech in an audio file located in a Google Cloud Storage
+                                                 bucket.
+  recognize.js sync-words <filename>             Detects speech in a local audio file with word time offset.
+  recognize.js async <filename>                  Creates a job to detect speech in a local audio file, and waits for the
+                                                 job to complete.
+  recognize.js async-gcs <gcsUri>                Creates a job to detect speech in an audio file located in a Google
+                                                 Cloud Storage bucket, and waits for the job to complete.
+  recognize.js async-gcs-words <gcsUri>          Creates a job to detect speech  with word time offset in an audio file
+                                                 located in a Google Cloud Storage bucket, and waits for the job to
+                                                 complete.
+  recognize.js stream <filename>                 Detects speech in a local audio file by streaming it to the Speech API.
+  recognize.js listen                            Detects speech in a microphone input stream. This command requires that
+                                                 you have SoX installed and available in your $PATH. See
+                                                 https://www.npmjs.com/package/node-record-lpcm16#dependencies
+  recognize.js sync-model <filename> <model>     Detects speech in a local audio file using provided model.
+  recognize.js sync-model-gcs <gcsUri> <model>   Detects speech in an audio file located in a Google Cloud Storage
+                                                 bucket using provided model.
+  recognize.js sync-auto-punctuation <filename>  Detects speech in a local audio file with auto punctuation.
+  recognize.js sync-enhanced-model <filename>    Detects speech in a local audio file using an enhanced model.
 
 Options:
   --version              Show version number                                                                   [boolean]
@@ -61,6 +68,10 @@ Examples:
   node recognize.js async-gcs gs://gcs-test-data/vr.flac -e FLAC -r 16000
   node recognize.js stream ./resources/audio.raw  -e LINEAR16 -r 16000
   node recognize.js listen
+  node recognize.js sync-model ./resources/Google_Gnome.wav video -e LINEAR16 -r 16000
+  node recognize.js sync-model-gcs gs://gcs-test-data/Google_Gnome.wav phone_call -e FLAC -r 16000
+  node recognize.js sync-auto-punctuation ./resources/commercial_mono.wav
+  node recognize.js sync-enhanced-model ./resources/commercial_mono.wav
 
 For more information, see https://cloud.google.com/speech/docs
 ```
@@ -80,12 +91,7 @@ __Usage:__ `node recognize.v1p1beta1.js --help`
 recognize.v1p1beta1.js <command>
 
 Commands:
-  recognize.v1p1beta1.js sync-model <filename> <model>     Detects speech in a local audio file using provided model.
-  recognize.v1p1beta1.js sync-model-gcs <gcsUri> <model>   Detects speech in an audio file located in a Google Cloud
-                                                           Storage bucket using provided model.
-  recognize.v1p1beta1.js sync-auto-punctuation <filename>  Detects speech in a local audio file with auto punctuation.
-  recognize.v1p1beta1.js sync-metadata <filename>          Detects speech in a local audio file with metadata.
-  recognize.v1p1beta1.js sync-enhanced-model <filename>    Detects speech in a local audio file using an enhanced model.
+  recognize.v1p1beta1.js sync-metadata <filename>  Detects speech in a local audio file with metadata.
 
 Options:
   --version              Show version number                                                                   [boolean]
@@ -95,11 +101,7 @@ Options:
   --help                 Show help                                                                             [boolean]
 
 Examples:
-  node recognize.v1p1beta1.js sync-model ./resources/Google_Gnome.wav video -e LINEAR16 -r 16000
-  node recognize.v1p1beta1.js sync-model-gcs gs://gcs-test-data/Google_Gnome.wav phone_call -e FLAC -r 16000
-  node recognize.v1p1beta1.js sync-auto-punctuation ./resources/commercial_mono.wav
   node recognize.v1p1beta1.js sync-metadata ./resources/commercial_mono.wav
-  node recognize.v1p1beta1.js sync-enhanced-model ./resources/commercial_mono.wav
 
 For more information, see https://cloud.google.com/speech/docs
 ```
