+++ samples/recognize.v1p1beta1.js
@@ -23,179 +23,6 @@
 
 'use strict';
 
-function syncRecognizeModelSelection(
-  filename,
-  model,
-  encoding,
-  sampleRateHertz,
-  languageCode
-) {
-  // [START speech_transcribe_model_selection_beta]
-  // Imports the Google Cloud client library for Beta API
-  /**
-   * TODO(developer): Update client library import to use new
-   * version of API when desired features become available
-   */
-  const speech = require('@google-cloud/speech').v1p1beta1;
-  const fs = require('fs');
-
-  // Creates a client
-  const client = new speech.SpeechClient();
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const filename = 'Local path to audio file, e.g. /path/to/audio.raw';
-  // const model = 'Model to use, e.g. phone_call, video, default';
-  // const encoding = 'Encoding of the audio file, e.g. LINEAR16';
-  // const sampleRateHertz = 16000;
-  // const languageCode = 'BCP-47 language code, e.g. en-US';
-
-  const config = {
-    encoding: encoding,
-    sampleRateHertz: sampleRateHertz,
-    languageCode: languageCode,
-    model: model,
-  };
-  const audio = {
-    content: fs.readFileSync(filename).toString('base64'),
-  };
-
-  const request = {
-    config: config,
-    audio: audio,
-  };
-
-  // Detects speech in the audio file
-  client
-    .recognize(request)
-    .then(data => {
-      const response = data[0];
-      const transcription = response.results
-        .map(result => result.alternatives[0].transcript)
-        .join('\n');
-      console.log(`Transcription: `, transcription);
-    })
-    .catch(err => {
-      console.error('ERROR:', err);
-    });
-  // [END speech_transcribe_model_selection_beta]
-}
-
-function syncRecognizeModelSelectionGCS(
-  gcsUri,
-  model,
-  encoding,
-  sampleRateHertz,
-  languageCode
-) {
-  // [START speech_transcribe_model_selection_gcs_beta]
-  // Imports the Google Cloud client library for Beta API
-  /**
-   * TODO(developer): Update client library import to use new
-   * version of API when desired features become available
-   */
-  const speech = require('@google-cloud/speech').v1p1beta1;
-
-  // Creates a client
-  const client = new speech.SpeechClient();
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const gcsUri = 'gs://my-bucket/audio.raw';
-  // const model = 'Model to use, e.g. phone_call, video, default';
-  // const encoding = 'Encoding of the audio file, e.g. LINEAR16';
-  // const sampleRateHertz = 16000;
-  // const languageCode = 'BCP-47 language code, e.g. en-US';
-
-  const config = {
-    encoding: encoding,
-    sampleRateHertz: sampleRateHertz,
-    languageCode: languageCode,
-    model: model,
-  };
-  const audio = {
-    uri: gcsUri,
-  };
-
-  const request = {
-    config: config,
-    audio: audio,
-  };
-
-  // Detects speech in the audio file
-  client
-    .recognize(request)
-    .then(data => {
-      const response = data[0];
-      const transcription = response.results
-        .map(result => result.alternatives[0].transcript)
-        .join('\n');
-      console.log(`Transcription: `, transcription);
-    })
-    .catch(err => {
-      console.error('ERROR:', err);
-    });
-  // [END speech_transcribe_model_selection_gcs_beta]
-}
-
-function syncRecognizeWithAutoPunctuation(
-  filename,
-  encoding,
-  sampleRateHertz,
-  languageCode
-) {
-  // [START speech_transcribe_auto_punctuation_beta]
-  // Imports the Google Cloud client library for Beta API
-  /**
-   * TODO(developer): Update client library import to use new
-   * version of API when desired features become available
-   */
-  const speech = require('@google-cloud/speech').v1p1beta1;
-  const fs = require('fs');
-
-  // Creates a client
-  const client = new speech.SpeechClient();
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const filename = 'Local path to audio file, e.g. /path/to/audio.raw';
-  // const encoding = 'Encoding of the audio file, e.g. LINEAR16';
-  // const sampleRateHertz = 16000;
-  // const languageCode = 'BCP-47 language code, e.g. en-US';
-
-  const config = {
-    encoding: encoding,
-    languageCode: languageCode,
-    enableAutomaticPunctuation: true,
-  };
-  const audio = {
-    content: fs.readFileSync(filename).toString('base64'),
-  };
-
-  const request = {
-    config: config,
-    audio: audio,
-  };
-
-  // Detects speech in the audio file
-  client
-    .recognize(request)
-    .then(data => {
-      const response = data[0];
-      const transcription = response.results
-        .map(result => result.alternatives[0].transcript)
-        .join('\n');
-      console.log(`Transcription: `, transcription);
-    })
-    .catch(err => {
-      console.error('ERROR:', err);
-    });
-  // [END speech_transcribe_auto_punctuation_beta]
-}
-
 function syncRecognizeWithMetaData(
   filename,
   encoding,
@@ -261,121 +88,14 @@ function syncRecognizeWithMetaData(
   // [END speech_transcribe_recognition_metadata_beta]
 }
 
-function syncRecognizeWithEnhancedModel(
-  filename,
-  encoding,
-  sampleRateHertz,
-  languageCode
-) {
-  // [START speech_transcribe_enhanced_model_beta]
-  // Imports the Google Cloud client library for Beta API
-  /**
-   * TODO(developer): Update client library import to use new
-   * version of API when desired features become available
-   */
-  const speech = require('@google-cloud/speech').v1p1beta1;
-  const fs = require('fs');
-
-  // Creates a client
-  const client = new speech.SpeechClient();
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const filename = 'Local path to audio file, e.g. /path/to/audio.raw';
-  // const encoding = 'Encoding of the audio file, e.g. LINEAR16';
-  // const sampleRateHertz = 16000;
-  // const languageCode = 'BCP-47 language code, e.g. en-US';
-
-  const config = {
-    encoding: encoding,
-    languageCode: languageCode,
-    useEnhanced: true,
-    model: 'phone_call',
-  };
-  const audio = {
-    content: fs.readFileSync(filename).toString('base64'),
-  };
-
-  const request = {
-    config: config,
-    audio: audio,
-  };
-
-  // Detects speech in the audio file
-  client
-    .recognize(request)
-    .then(data => {
-      const response = data[0];
-      response.results.forEach(result => {
-        const alternative = result.alternatives[0];
-        console.log(alternative.transcript);
-      });
-    })
-    .catch(err => {
-      console.error('ERROR:', err);
-    });
-  // [END speech_transcribe_enhanced_model_beta]
-}
-
 require(`yargs`)
   .demand(1)
-  .command(
-    `sync-model <filename> <model>`,
-    `Detects speech in a local audio file using provided model.`,
-    {},
-    opts =>
-      syncRecognizeModelSelection(
-        opts.filename,
-        opts.model,
-        opts.encoding,
-        opts.sampleRateHertz,
-        opts.languageCode
-      )
-  )
-  .command(
-    `sync-model-gcs <gcsUri> <model>`,
-    `Detects speech in an audio file located in a Google Cloud Storage bucket using provided model.`,
-    {},
-    opts =>
-      syncRecognizeModelSelectionGCS(
-        opts.gcsUri,
-        opts.model,
-        opts.encoding,
-        opts.sampleRateHertz,
-        opts.languageCode
-      )
-  )
-  .command(
-    `sync-auto-punctuation <filename>`,
-    `Detects speech in a local audio file with auto punctuation.`,
-    {},
-    opts =>
-      syncRecognizeWithAutoPunctuation(
-        opts.filename,
-        opts.encoding,
-        opts.sampleRateHertz,
-        opts.languageCode
-      )
-  )
   .command(
     `sync-metadata <filename>`,
     `Detects speech in a local audio file with metadata.`,
     {},
     opts =>
       syncRecognizeWithMetaData(
-        opts.filename,
-        opts.encoding,
-        opts.sampleRateHertz,
-        opts.languageCode
-      )
-  )
-  .command(
-    `sync-enhanced-model <filename>`,
-    `Detects speech in a local audio file using an enhanced model.`,
-    {},
-    opts =>
-      syncRecognizeWithEnhancedModel(
         opts.filename,
         opts.encoding,
         opts.sampleRateHertz,
@@ -405,15 +125,7 @@ require(`yargs`)
       type: 'string',
     },
   })
-  .example(
-    `node $0 sync-model ./resources/Google_Gnome.wav video -e LINEAR16 -r 16000`
-  )
-  .example(
-    `node $0 sync-model-gcs gs://gcs-test-data/Google_Gnome.wav phone_call -e FLAC -r 16000`
-  )
-  .example(`node $0 sync-auto-punctuation ./resources/commercial_mono.wav`)
   .example(`node $0 sync-metadata ./resources/commercial_mono.wav`)
-  .example(`node $0 sync-enhanced-model ./resources/commercial_mono.wav`)
   .wrap(120)
   .recommendCommands()
   .epilogue(`For more information, see https://cloud.google.com/speech/docs`)
