+++ samples/system-test/recognize.test.js
@@ -13,8 +13,6 @@
  * limitations under the License.
  */
 
-/* eslint-disable */
-
 'use strict';
 
 const path = require(`path`);
@@ -29,8 +27,15 @@ const bucketName = `nodejs-docs-samples-test-${uuid.v4()}`;
 const cmd = `node recognize.js`;
 const cwd = path.join(__dirname, `..`);
 const filename = `audio.raw`;
+const filename1 = `Google_Gnome.wav`;
+const filename2 = `commercial_mono.wav`;
 const filepath = path.join(__dirname, `../resources/${filename}`);
+const filepath1 = path.join(__dirname, `../resources/${filename1}`);
+const filepath2 = path.join(__dirname, `../resources/${filename2}`);
 const text = `how old is the Brooklyn Bridge`;
+const text1 = `the weather outside is sunny`;
+const text2 = `Terrific. It's on the way.`;
+const text3 = `Chrome`;
 
 test.before(async () => {
   const [bucket] = await storage.createBucket(bucketName);
@@ -89,4 +94,34 @@ test(`should run async recognize on a GCS file with word time offset`, async t =
 test(`should run streaming recognize`, async t => {
   const output = await runAsync(`${cmd} stream ${filepath}`, cwd);
   t.true(output.includes(`Transcription: ${text}`));
+});
+
+test(`should run sync recognize with model selection`, async t => {
+  const model = `video`;
+  const output = await runAsync(`${cmd} sync-model ${filepath1} ${model}`, cwd);
+  t.true(output.includes(`Transcription:`));
+  t.true(output.includes(text1));
+});
+
+test(`should run sync recognize on a GCS file with model selection`, async t => {
+  const model = `video`;
+  const output = await runAsync(
+    `${cmd} sync-model-gcs gs://${bucketName}/${filename1} ${model}`,
+    cwd
+  );
+  t.true(output.includes(`Transcription:`));
+  t.true(output.includes(text1));
+});
+
+test(`should run sync recognize with auto punctuation`, async t => {
+  const output = await runAsync(
+    `${cmd} sync-auto-punctuation ${filepath2}`,
+    cwd
+  );
+  t.true(output.includes(text2));
+});
+
+test(`should run sync recognize with enhanced model`, async t => {
+  const output = await runAsync(`${cmd} sync-enhanced-model ${filepath2}`, cwd);
+  t.true(output.includes(text3));
 });
