+++ samples/system-test/recognize.v1p1beta1.test.js
@@ -13,17 +13,13 @@
  * limitations under the License.
  */
 
-/* eslint-disable */
-
 'use strict';
 
 const path = require(`path`);
 const {Storage} = require(`@google-cloud/storage`);
 const test = require(`ava`);
 const uuid = require(`uuid`);
-
 const {runAsync} = require(`@google-cloud/nodejs-repo-tools`);
-
 const storage = new Storage();
 const bucketName = `nodejs-docs-samples-test-${uuid.v4()}`;
 const cmd = `node recognize.v1p1beta1.js`;
@@ -32,9 +28,8 @@ const filename1 = `Google_Gnome.wav`;
 const filename2 = `commercial_mono.wav`;
 const filepath1 = path.join(__dirname, `../resources/${filename1}`);
 const filepath2 = path.join(__dirname, `../resources/${filename2}`);
-const text1 = `the weather outside is sunny`;
-const text2 = `Terrific. It's on the way.`;
-const text3 = `Chrome`;
+
+const text = `Chrome`;
 
 test.before(async () => {
   const [bucket] = await storage.createBucket(bucketName);
@@ -48,37 +43,7 @@ test.after.always(async () => {
   await bucket.delete();
 });
 
-test(`should run sync recognize with model selection`, async t => {
-  const model = `video`;
-  const output = await runAsync(`${cmd} sync-model ${filepath1} ${model}`, cwd);
-  t.true(output.includes(`Transcription:`));
-  t.true(output.includes(text1));
-});
-
-test(`should run sync recognize on a GCS file with model selection`, async t => {
-  const model = `video`;
-  const output = await runAsync(
-    `${cmd} sync-model-gcs gs://${bucketName}/${filename1} ${model}`,
-    cwd
-  );
-  t.true(output.includes(`Transcription:`));
-  t.true(output.includes(text1));
-});
-
-test(`should run sync recognize with auto punctuation`, async t => {
-  const output = await runAsync(
-    `${cmd} sync-auto-punctuation ${filepath2}`,
-    cwd
-  );
-  t.true(output.includes(text2));
-});
-
 test(`should run sync recognize with metadata`, async t => {
   const output = await runAsync(`${cmd} sync-metadata ${filepath2}`, cwd);
-  t.true(output.includes(text3));
-});
-
-test(`should run sync recognize with enhanced model`, async t => {
-  const output = await runAsync(`${cmd} sync-enhanced-model ${filepath2}`, cwd);
-  t.true(output.includes(text3));
+  t.true(output.includes(text));
 });
